requirements.txt
pytest==8.3.3
sentry-sdk==2.16.0
tenacity==9.0.0
requests==2.32.3
yfinance==0.2.44
types-requests
types-pymysql
types-pygments
types-colorama
types-setuptools
pandas-stubs